Use Types::Standard instead of Dancer2::Core::Types.
Type::Tiny is now in the sad state that it has no active maintainer and
is broken in recent Perls.

https://rt.cpan.org/Public/Bug/Display.html?id=114386

Because of this problem, and after careful deliberation, the Dancer2 core
team decided to switch back to MooX::Types::MooseLike for the next
release. Rather than chasing Dancer2 we just stop using
Dancer2::Core::Types and switch back to Types::Standard.